create named files, simplify project structure
